requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pytest==7.4.3
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
pyrfc==3.0.5  # For SAP RFC communication
schedule==1.2.1
aiohttp==3.9.1
prometheus-client==0.19.0 